refactor: Form 클래스 구현
1. CarNameForm 클래스와 RacingCountForm 클래스가 Form 클래스를 상속받도록 구현
2. CarNameForm 클래스와 RacingCountForm 클래스에서 공통된 메서드를 Form 클래스로 이동
- Form 클래스로 이동시킨 메서드 목록
    - addSubmitClickEventHandler
    - handleException
